# backend/src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API",
      "description": "Backend API Information",
      "version": "1.0"
    },
    "securityDefinitions": {
      "bearerAuth": {
        "name": "Authorization",
        "in": "header",
        "type": "apiKey",
        "description": "JWT Authorization header"
      }
    },
    "security": [{ "bearerAuth": []}],
...
    "paths": {
...
        "/api/auth/connected": {
           "get": {
              "tags": ["Authentification"],
              "x-swagger-router-controller": "auth",
              "operationId": "connected",
              "parameters": [],
              "responses": {
                "default": {
                  "description": "Unexpected error",
                  "schema": {
                    "$ref": "#/definitions/Error"
                  }
                },
                "200":{
                  "description":"Successful request.",
                  "schema":{
                     "$ref":"#/definitions/getUser"
                  }
               }
              }
           }
        },
...
        "/api/purchases/{id}":{
          "get":{
             "tags":["Purchases"],
             "x-swagger-router-controller":"purchases",
             "operationId":"getPurchaseById",
             "parameters": [
               {
                  "name": "id",
                  "in":"path",
                  "type": "string",
                  "required": true
               }
             ],
             "responses":{
                "default":{
                   "description":"Unexpected error",
                   "schema":{
                      "$ref":"#/definitions/Error"
                   }
                },
                "200":{
                   "description":"Successful request.",
                   "schema":{
                      "$ref":"#/definitions/getPurchase"
                   }
                }
             }
          },
          "delete":{
             "tags":["Purchases"],
             "x-swagger-router-controller":"purchases",
             "operationId":"deletePurchase",
             "parameters": [
               {
                  "name": "id",
                  "in":"path",
                  "type": "string",
                  "required": true
               }
             ],
             "responses":{
                "default":{
                   "description":"Unexpected error",
                   "schema":{
                      "$ref":"#/definitions/Error"
                   }
                },
                "201":{
                   "description":"Successful request.",
                   "schema":{
                      "$ref":"#/definitions/Success"
                   }
                }
             }
          }
        }
     },
    "definitions": {
        "Error": {
           "required": ["error"],
           "properties": {
              
              "error": {
                 "type": "string"
              }
           }
        },
        "Success": {
           "required": ["message"],
           "properties": {
             
              "message": {
                 "type": "string"
              }
           }
        },
...
        "getUser": {
           "type": "object",
           "properties": {
              "id": { "type": "string" },
              "firstname": { "type": "string" },
              "lastname": { "type": "string" },
              "email": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
           }
        },
...
         "getPurchase": {
            "type": "object",
            "properties": {
               "id": { "type": "string" },
               "createdAt": { "type": "string" },
               "updatedAt": { "type": "string" },
               "product": {
                  "type": "object",
                  "properties": {
                     "id": { "type": "string" },
                     "label": { "type": "string" },
                     "price": { "type": "integer" },
                     "desc": { "type": "string" },
                     "picture": { "type": "string" },
                     "createdAt": { "type": "string" },
                     "updatedAt": { "type": "string" },
                     "products_tags": {
                        "type": "object",
                        "properties": {
                           "tag": {
                              "type": "object",
                              "properties": {
                                 "label": { "type": "string" }
                              }
                           }
                        }
                     }
                  }
               },
               "user": {
                  "type": "object",
                  "properties": {
                     "id": { "type": "string" },
                     "firstname": { "type": "string" },
                     "lastname": { "type": "string" },
                     "email": { "type": "string" },
                     "createdAt": { "type": "string" },
                     "updatedAt": { "type": "string" }
                  }
               }
            }
         },
...
     }
  }